@@ -1,6 +1,6 @@
 {
   "name": "capra_web_ui",
-  "version": "2.3.16",
+  "version": "2.3.19",
   "author": {
     "name": "Club Capra",
     "email": "admin@clubcapra.com",
@@ -19,10 +19,9 @@
     "typecheck": "tsc --noEmit",
     "clean": "electron-snowpack clean",
     "postinstall": "patch-package",
-    "depcheck": "depcheck --ignores=\"ts-node,ts-jest,@jest/types,@types/jest\"",
-    "audit": "better-npm-audit audit",
+    "depcheck": "depcheck --ignores=\"ts-node,ts-jest,@jest/types,@types/jest,concurrently\"",
     "ci": "concurrently --kill-others-on-fail npm:lint npm:typecheck npm:test",
-    "dependencies": "concurrently --kill-others-on-fail npm:depcheck npm:audit",
+    "dependencies": "npm run depcheck",
     "local_ci": "concurrently npm:ci npm:dependencies"
   },
   "build": {
@@ -47,56 +46,55 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.6.2",
-    "@xstate/react": "^1.6.1",
-    "better-npm-audit": "^3.4.0",
+    "@xstate/react": "^1.6.2",
     "chalk": "^4.1.2",
-    "chart.js": "^3.5.1",
+    "chart.js": "^3.6.0",
     "chartjs-adapter-date-fns": "^2.0.0",
-    "concurrently": "^6.3.0",
-    "date-fns": "^2.25.0",
+    "date-fns": "^2.26.0",
     "electron-log": "^4.4.1",
-    "execa": "^5.1.1",
+    "execa": "^6.0.0",
     "lodash": "^4.17.21",
-    "nanoid": "^3.1.29",
+    "nanoid": "^3.1.30",
     "polished": "^4.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
-    "react-redux": "^7.2.5",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "react-toastify": "^8.0.3",
+    "react-redux": "^7.2.6",
+    "react-router": "^6.0.2",
+    "react-router-dom": "^6.0.2",
+    "react-toastify": "^8.1.0",
     "ros3d": "^1.0.1",
     "roslib": "^1.1.0",
-    "styled-components": "^5.3.1",
-    "xstate": "^4.25.0"
+    "styled-components": "^5.3.3",
+    "xstate": "^4.26.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.2",
-    "@types/lodash": "^4.14.175",
-    "@types/react": "^17.0.27",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.19",
-    "@types/react-router-dom": "^5.3.1",
-    "@types/roslib": "^1.1.3",
+    "@types/jest": "^27.0.3",
+    "@types/lodash": "^4.14.177",
+    "@types/react": "^17.0.36",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/roslib": "^1.1.4",
     "@types/styled-components": "^5.1.15",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "concurrently": "^6.4.0",
     "depcheck": "^1.4.2",
-    "electron": "^15.1.2",
-    "electron-builder": "^22.13.1",
+    "electron": "^16.0.1",
+    "electron-builder": "^22.14.5",
     "electron-devtools-installer": "^3.2.0",
     "electron-snowpack": "^0.12.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.3.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "^27.2.5",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "jest": "^27.3.1",
     "patch-package": "^6.4.7",
     "prettier": "^2.4.1",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.3"
+    "ts-jest": "^27.0.7",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.2"
   }
 }
