@@ -1,26 +1,19 @@
-import {
-  AUDIO_MSG,
-  AUDIO_MSG_TYPE,
-  AUDIO_START,
-  AUDIO_STOP,
-} from '@/shared/constants'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
-const { ipcRenderer } = window.require('electron')
 
 export const useAudio = () => {
   useEffect(() => {
-    ipcRenderer.on(AUDIO_MSG, (_event, { error }: AUDIO_MSG_TYPE) => {
+    window.preloadApi.audio.receive(({ error }) => {
       if (error) {
         toast.error(error)
       }
     })
   }, [])
   const start = () => {
-    ipcRenderer.send(AUDIO_START)
+    window.preloadApi.audio.start()
   }
   const stop = () => {
-    ipcRenderer.send(AUDIO_STOP)
+    window.preloadApi.audio.stop()
   }
   return [start, stop]
 }
